% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_df_to_matrix}
\alias{convert_df_to_matrix}
\title{Convert a dataframe to a matrix for chord diagrams}
\usage{
convert_df_to_matrix(df, column_var, cell)
}
\arguments{
\item{df}{A data frame with three columns: a row variable, a column variable,
and a value variable.}

\item{column_var}{The name of the variable to use for columns.}

\item{cell}{The name of the variable to use for filling the matrix.}
}
\value{
A square matrix.
}
\description{
Convert a dataframe to a matrix for chord diagrams
}
