% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_plot.R
\name{plot_time_use}
\alias{plot_time_use}
\title{Plot area plot of time use.}
\usage{
plot_time_use(tbl, x_max = 1440)
}
\arguments{
\item{tbl}{tbl of data with at least the following columns: \code{per_id},
\code{event_type}, \code{start}, \code{end}.}

\item{x_max}{Maximum time on x-axis.}
}
\value{
A plot.
}
\description{
Assumes \code{start} and \code{end} in \code{tbl} and the \code{x_max} value
are in simulation time (minutes from 4:00am) with \code{end} missing for the
last activity of every person.
}

