% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{plot_trips_per_person}
\alias{plot_trips_per_person}
\title{Plot histogram of trips per person.}
\usage{
plot_trips_per_person(tbl)
}
\arguments{
\item{tbl}{tbl of data with at least the following columns: .}
}
\value{
A plot.
}
\description{
Plot histogram of trips per person.
}
